@@ -4,67 +4,46 @@
   "private": true,
   "homepage": ".",
   "dependencies": {
-    "@bcgov/bootstrap-theme": "https://github.com/bcgov/bootstrap-theme/releases/download/v1.1.1/bcgov-bootstrap-theme-1.1.1.tgz",
     "@beam-australia/react-env": "^3.1.1",
-    "@types/jest": "^24.0.23",
-    "@types/moment-timezone": "^0.5.13",
-    "@types/node": "^12.12.16",
-    "@types/react": "^16.9.16",
+    "@types/node": "^16.18.10",
+    "@types/react": "^18.0.26",
     "@types/react-csv": "^1.1.1",
     "@types/react-datepicker": "^2.11.0",
-    "@types/react-dom": "^16.9.4",
-    "@types/react-redux": "^7.1.7",
+    "@types/react-dom": "^18.0.9",
     "@types/react-router": "^5.1.5",
     "@types/react-router-dom": "^5.1.3",
     "@types/react-select": "^3.0.13",
     "@types/react-table": "^7.0.16",
-    "@typescript-eslint/eslint-plugin": "^2.11.0",
-    "@typescript-eslint/parser": "^2.11.0",
-    "babel-eslint": "10.0.1",
-    "bootstrap": "^4.4.1",
-    "class-transformer": "^0.2.3",
+    "bootstrap": "~4.3.1",
+    "class-transformer": "^0.5.1",
     "client-oauth2": "^4.2.5",
     "email-validator": "^2.0.4",
-    "eslint-config-prettier": "^6.7.0",
-    "eslint-config-react": "^1.1.7",
-    "eslint-plugin-prettier": "^3.1.1",
     "jquery": "^3.4.1",
-    "js-levenshtein": "^1.1.6",
-    "merge": "^1.2.1",
+    "keycloak-js": "^21.0.0",
     "moment": "^2.26.0",
-    "moment-timezone": "^0.5.31",
-    "oidc-client": "^1.9.0",
-    "prettier": "^1.19.1",
+    "moment-timezone": "^0.5.41",
     "query-string": "^6.12.1",
-    "react": "^16.0.0",
+    "react": "^18.2.0",
     "react-csv": "^2.0.3",
-    "react-datepicker": "^2.16.0",
-    "react-dom": "^16.0.0",
-    "react-redux": "^7.2.0",
-    "react-router-bootstrap": "^0.24.4",
+    "react-datepicker": "3",
+    "react-dom": "^18.2.0",
     "react-router-dom": "^5.1.2",
-    "react-scripts": "^3.3.1",
-    "react-select": "^3.1.0",
+    "react-scripts": "5.0.1",
+    "react-select": "^5.7.0",
     "react-table": "^7.0.4",
-    "redux-oidc": "^4.0.0-beta1",
-    "redux-persist": "^6.0.0",
     "reflect-metadata": "^0.1.13",
     "rimraf": "^2.6.2",
-    "sass": "^1.56.2",
+    "sass": "^1.57.1",
     "slugify": "^1.4.0",
-    "typescript-eslint": "^0.0.1-alpha.0"
+    "typescript": "^4.9.4",
+    "web-vitals": "^2.1.4"
   },
   "devDependencies": {
-    "ajv": "^6.9.1",
-    "cross-env": "^5.2.0",
-    "eslint": "^5.12.0",
-    "eslint-config-react-app": "^4.0.1",
-    "eslint-plugin-flowtype": "^2.0.0",
-    "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-jsx-a11y": "^6.2.1",
-    "eslint-plugin-react": "^7.11.1",
-    "eslint-plugin-react-hooks": "^4.0.4",
-    "typescript": "^3.7.3"
+    "@bcgov/bootstrap-theme": "https://github.com/bcgov/bootstrap-theme/releases/download/v1.1.3/bcgov-bootstrap-theme-1.1.3.tgz",
+    "@types/jest": "^29.4.0",
+    "eslint-config-react": "^1.1.7",
+    "eslint-plugin-prettier": "^3.1.1",
+    "prettier": "^2.8.1"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -73,9 +52,7 @@
     "prestart": "react-env --dest=public/config",
     "start": "rimraf ./build && react-scripts start",
     "build": "react-scripts build",
-    "test": "cross-env CI=true react-scripts test --env=jsdom",
-    "eject": "react-scripts eject",
-    "lint": "eslint ./src/"
+    "eject": "react-scripts eject"
   },
   "browserslist": {
     "production": [
