@@ -1,30 +1,22 @@
-import React, { useEffect } from 'react'
+import React from 'react'
 
-import { signinRedirectOptions } from '../../helpers/envHelper'
-import userManager from '../../store/utils/userManager'
+import FAIcon from '../DisplayHelpers/Interface/Icons/FAIcon'
 
 const Unauthorized = (): JSX.Element => {
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      userManager.signinRedirect(signinRedirectOptions)
-    }, 100)
-    return (): void => clearTimeout(timer)
-  }, [])
-
   return (
     <div className="Centered row">
       <div className="col-6 offset-3">
         <h1 className="text-primary display-4 my-5">New Job Survey Admin</h1>
-        <h1>You must log in to view this page.</h1>
-        <p>You will be redirected shortly. If not, click the button below.</p>
-        <button
-          className="btn btn-lg btn-outline-primary mt-3"
-          onClick={(): Promise<void> =>
-            userManager.signinRedirect(signinRedirectOptions)
-          }
-        >
-          Click here to log in &rarr;
-        </button>
+        <h1 className="mb-4">
+          <FAIcon name="exclamation-circle text-danger" /> Not authorized
+        </h1>
+        <p className="lead">
+          You successfully logged in via Keycloak, but you do not have the
+          correct permissions to use this application.
+        </p>
+        <p className="lead">
+          Please contact the BC Stats team to obtain access.
+        </p>
       </div>
     </div>
   )
