@@ -1,26 +1,22 @@
 import React from 'react'
-import { connect } from 'react-redux'
 
-import { FixTypeLater } from '../../types/FixTypeLater'
-import { mapUserToPropsFromState } from '../../helpers/userHelper'
+import { authRole } from '../../helpers/envHelper'
+import KeycloakService from '../Login/KeycloakService'
+import Unauthenticated from '../Login/Unauthenticated'
 import Unauthorized from '../Login/Unauthorized'
 
-interface OwnProps {
+interface AuthWrapperProps {
   children: React.ReactNode
 }
 
-interface StateProps {
-  user: FixTypeLater
+const AuthWrapper = (props: AuthWrapperProps): JSX.Element => {
+  const isLoggedIn = KeycloakService.isLoggedIn()
+  const hasCorrectRole = KeycloakService.hasRole([authRole()])
+
+  if (!isLoggedIn) return <Unauthenticated />
+  if (!hasCorrectRole) return <Unauthorized />
+
+  return <>{props.children}</>
 }
 
-interface Props extends OwnProps, StateProps {}
-
-class AuthWrapper extends React.Component<Props> {
-  render(): React.ReactNode {
-    const { user } = this.props
-
-    return !user || user.expired ? <Unauthorized /> : this.props.children
-  }
-}
-
-export default connect(mapUserToPropsFromState)(AuthWrapper)
+export default AuthWrapper
