@@ -1,5 +1,5 @@
+import * as React from 'react'
 import { Route } from 'react-router-dom'
-import React from 'react'
 
 import { FixTypeLater } from '../../types/FixTypeLater'
 import AuthWrapper from './AuthWrapper'
