@@ -7,10 +7,10 @@ export class AdminSetting {
   displayName?: string
   value?: string
 
-  @Transform((date: string) => dateOrUndefined(date, true))
+  @Transform(({ value }: { value: string }) => dateOrUndefined(value, true))
   createdTs?: Date
 
-  @Transform((date: string) => dateOrUndefined(date, true))
+  @Transform(({ value }: { value: string }) => dateOrUndefined(value, true))
   modifiedTs?: Date
 }
 
