@@ -9,9 +9,9 @@ export class EmployeeTimelineEntry {
   comment?: string
   adminUserName?: string
 
-  @Transform((date: string) => dateOrUndefined(date, true))
+  @Transform(({ value }: { value: string }) => dateOrUndefined(value, true))
   createdTs?: Date
 
-  @Transform((date: string) => dateOrUndefined(date, true))
+  @Transform(({ value }: { value: string }) => dateOrUndefined(value, true))
   modifiedTs?: Date
 }
