@@ -1,6 +1,8 @@
-using System;
-using NewJobSurveyAdmin.Models;
 using Microsoft.Extensions.Options;
+using NewJobSurveyAdmin.Models;
+using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Net.Http;
 using System.Threading.Tasks;
 
@@ -24,38 +26,161 @@ namespace NewJobSurveyAdmin.Services.CallWeb
             );
         }
 
-        // Determines whether a survey is complete, given a telkey.
-        public async Task<string> GetSurveyStatusCode(Employee employee)
+        // Determines whether a survey is complete, given multiple employees.
+        public async Task<TaskResult<Tuple<Employee, string>>> GetSurveyStatusCodes(
+            List<Employee> employees
+        )
         {
-            var callWebDto = await CallWebApi.GetOne(employee.Telkey);
+            var taskResult = new TaskResult<Tuple<Employee, string>>();
 
-            if (callWebDto.IsSurveyComplete.Equals("1"))
+            try
             {
-                return EmployeeStatusEnum.SurveyComplete.Code;
+                var telkeys = employees.Select(e => e.Telkey).ToArray();
+
+                if (telkeys.Count() != employees.Count())
+                {
+                    throw new Exception("telkeys.Count != employees.Count!");
+                }
+
+                var results = (await CallWebApi.GetMultiple(telkeys)).ToList();
+
+                foreach (var employee in employees)
+                {
+                    var result = results.Find(dto => dto.Telkey == employee.Telkey);
+
+                    if (result == null)
+                    {
+                        taskResult.AddFailedWithException(
+                            new Tuple<Employee, string>(employee, null),
+                            new CallWebRetrieveFailedException(
+                                $"No GetSurveyStatusCodes result for {employee}"
+                            )
+                        );
+                    }
+                    else
+                    {
+                        var statusCode =
+                            (IsSurveyComplete(result))
+                                ? EmployeeStatusEnum.SurveyComplete.Code
+                                : result.CurrentStatus;
+                        taskResult.AddSucceeded(new Tuple<Employee, string>(employee, statusCode));
+                    }
+                }
+            }
+            catch (Exception exception)
+            {
+                // Assume the entire operation failed.
+                foreach (var employee in employees)
+                {
+                    taskResult.AddFailed(new Tuple<Employee, string>(employee, null));
+                }
+                taskResult.AddException(
+                    new CallWebRetrieveFailedException(
+                        $"GetSurveyStatusCodes failed for a range of employees: {String.Join(", ", employees)}. Error: {exception.Message}"
+                    )
+                );
             }
 
-            return callWebDto.CurrentStatus;
+            return taskResult;
         }
 
-        public async Task<string> CreateSurvey(Employee employee)
+        public async Task<TaskResult<Employee>> CreateSurveys(List<Employee> employees)
         {
-            var callWebPostDto = CallWebPostDto.FromEmployee(employee);
-            var callWebDto = await CallWebApi.Post(callWebPostDto);
+            var taskResult = new TaskResult<Employee>();
 
-            if (string.IsNullOrEmpty(callWebDto.Telkey))
+            try
             {
-                throw new InvalidOperationException("Telkey was not created.");
+                var callWebPostDtos = employees
+                    .Select(e => CallWebPostDto.FromEmployee(e))
+                    .ToList();
+
+                var results = (await CallWebApi.PostMultiple(callWebPostDtos)).ToList();
+
+                foreach (var employee in employees)
+                {
+                    // We can't compare telkeys because the employes we're
+                    // trying to create don't have them yet. Instead, compare on
+                    // first name, last name, and ministry.
+                    var result = results.Find(
+                        result =>
+                            employee.PreferredFirstName.Equals(result.PreferredFirstName)
+                            && employee.LastName.Equals(result.LastName)
+                    // && employee.Ministry.Equals(result.Ministry)
+                    // TODO: No Ministry on these employees. Any other field?
+                    );
+
+                    if (result == null)
+                    {
+                        taskResult.AddFailedWithException(
+                            employee,
+                            new CallWebNoTelkeyException(
+                                $"No telkey in CreateSurveys result for {employee}"
+                            )
+                        );
+                    }
+                    else
+                    {
+                        employee.Telkey = result.Telkey;
+                        taskResult.AddSucceeded(employee);
+                    }
+                }
+            }
+            catch (Exception exception)
+            {
+                // Assume the entire operation failed.
+                taskResult.AddFailed(employees);
+                taskResult.AddException(
+                    new CallWebCreateFailedException(
+                        $"CreateSurveys failed for a range of employees: {String.Join(", ", employees)}. Error: {exception.Message}"
+                    )
+                );
             }
 
-            return callWebDto.Telkey;
+            return taskResult;
         }
 
-        public async Task<CallWebRowDto> UpdateSurvey(Employee employee)
+        public async Task<TaskResult<Employee>> UpdateSurveys(List<Employee> employees)
         {
-            var callWebPatchDto = CallWebPatchDto.FromEmployee(employee);
-            var callWebDto = await CallWebApi.Patch(callWebPatchDto);
+            var taskResult = new TaskResult<Employee>();
 
-            return callWebDto;
+            try
+            {
+                var callWebPatchDtos = employees
+                    .Select(e => CallWebPatchDto.FromEmployee(e))
+                    .ToList();
+                var results = (await CallWebApi.PatchMultiple(callWebPatchDtos)).ToList();
+
+                foreach (var employee in employees)
+                {
+                    var result = results.Find(result => employee.Telkey.Equals(result.Telkey));
+
+                    if (result == null)
+                    {
+                        taskResult.AddFailedWithException(
+                            employee,
+                            new CallWebUpdateFailedException(
+                                $"No UpdateSurveys result for {employee}"
+                            )
+                        );
+                    }
+                    else
+                    {
+                        taskResult.AddSucceeded(employee);
+                    }
+                }
+            }
+            catch (Exception exception)
+            {
+                // Assume the entire operation failed.
+                taskResult.AddFailed(employees);
+                taskResult.AddException(
+                    new CallWebUpdateFailedException(
+                        $"UpdateSurveys failed for a range of employees: {String.Join(", ", employees)}. Error: {exception.Message}"
+                    )
+                );
+            }
+
+            return taskResult;
         }
 
         public async Task<CallWebRowDto[]> ListAll()
@@ -63,6 +188,11 @@ namespace NewJobSurveyAdmin.Services.CallWeb
             var response = await CallWebApi.GetAll();
 
             return response;
+        }
+
+        private bool IsSurveyComplete(CallWebRowDto callWebDto)
+        {
+            return callWebDto.IsSurveyComplete != null && callWebDto.IsSurveyComplete.Equals("1");
         }
     }
 }
